requirements.txt
flask==2.3.2
flask-cors==3.0.10
numpy==1.25.2
faiss-cpu==1.7.4
google-generativeai==0.1.0
python-dotenv==1.0.0
gunicorn==23.0.0